refactor: refactor database query classes from service to repository
