+++ tests/Unit/PasswordResetServiceTest.php
@@ -2,7 +2,7 @@
 
 namespace Tests\Unit;
 
-use App\Components\Services\PasswordResetService;
+use App\Components\Repositorys\PasswordResetRepository;
 use App\Components\TokenGenerator;
 use App\PasswordReset;
 use PHPUnit\Framework\TestCase;
@@ -14,39 +14,39 @@ class PasswordResetServiceTest extends TestCase
 {
     use MockeryPHPUnitIntegration;
 
-    public function makePasswordResetService($model = null, $tokenGenerator = null)
+    public function makePasswordResetRepository($model = null, $tokenGenerator = null)
     {
         $model = $model ? $model : new PasswordReset();
         $tokenGenerator = $tokenGenerator ? $tokenGenerator : new TokenGenerator;
-        return new PasswordResetService($model, $tokenGenerator);
+        return new PasswordResetRepository($model, $tokenGenerator);
     }
 
     /** @test */
-    public function password_reset_service_query_by_email_should_return_query_builder()
+    public function password_reset_repository_query_by_email_should_return_query_builder()
     {
         $model = Mockery::mock(PasswordReset::class);
         $model->shouldReceive('where->count')->andReturn(1);
 
-        $service = $this->makePasswordResetService($model);
+        $service = $this->makePasswordResetRepository($model);
         $queryBuilder = $service->queryByEmail('valid_email');
 
         $this->assertGreaterThan(0, $queryBuilder->count());
     }
 
     /** @test */
-    public function password_reset_service_find_by_email_should_return_null()
+    public function password_reset_repository_find_by_email_should_return_null()
     {
         $model = Mockery::mock(PasswordReset::class);
         $model->shouldReceive('where->first')->andReturn(null);
 
-        $service = $this->makePasswordResetService($model);
+        $service = $this->makePasswordResetRepository($model);
         $passwordReset = $service->findByEmail('invalid_email');
 
         $this->assertEquals(null, $passwordReset);
     }
 
     /** @test */
-    public function password_reset_service_find_by_email_should_return_password_reset()
+    public function password_reset_repository_find_by_email_should_return_password_reset()
     {
         $reset = new PasswordReset();
         $reset->token = 'valid_token';
@@ -55,7 +55,7 @@ class PasswordResetServiceTest extends TestCase
         $model = Mockery::mock(PasswordReset::class);
         $model->shouldReceive('where->first')->andReturn($reset);
 
-        $service = $this->makePasswordResetService($model);
+        $service = $this->makePasswordResetRepository($model);
         $passwordReset = $service->findByEmail($reset->email);
 
         $this->assertEquals($reset->email, $passwordReset->email);
@@ -63,7 +63,7 @@ class PasswordResetServiceTest extends TestCase
     }
 
     /** @test */
-    public function password_reset_service_should_store_reset_request()
+    public function password_reset_repository_should_store_reset_request()
     {
         $fields = ['email' => 'valid_mail@mail.com', 'token' => 'valid_token'];
 
@@ -74,7 +74,7 @@ class PasswordResetServiceTest extends TestCase
         $model->shouldReceive('fill')->with($fields);
         $model->shouldReceive('save')->andReturn(true);
 
-        $sut = $this->makePasswordResetService($model, $tokenGenerator);
+        $sut = $this->makePasswordResetRepository($model, $tokenGenerator);
 
         $this->assertTrue($sut->storePasswordResetRequest($fields['email'], $fields['token']));
     }
